feat: rename component name
